requirements.txt
annotated-types==0.7.0
azure-core==1.32.0
azure-cosmos==4.9.0
azure-identity==1.19.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.9.0
azure-keyvault-keys==4.10.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.24.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
eventlet==0.38.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msal==1.31.1
msal-extensions==1.2.0
portalocker==2.10.1
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
pyparsing==3.2.0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
requests==2.32.3
rsa==4.9
simple-websocket==1.1.0
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
